func do registro

+++ base/views.py
@@ -1,16 +1,22 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
-from .models import Room, Topic
-from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .forms import RoomForm
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
 from django.db.models import Q
+from .models import Room, Topic
+from .forms import RoomForm
 
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         
         try:
@@ -26,13 +32,30 @@ def loginPage(request):
         else:
             messages.error(request, 'Senha incorreta')
             
-    contexto = {}
+    contexto = {'page': page}
     return render(request, 'base/login_register.html', contexto)
 
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Aconteceu um erro durante o cadastro')
+            
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def home(request):
@@ -54,6 +77,7 @@ def room(request, pk):
     return render(request, 'base/room.html', contexto)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -66,9 +90,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', contexto)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse("Você não pode modificar esta sala!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -78,8 +107,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', contexto)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("Você não pode modificar esta sala!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
